feat: 🎸 Added Blueprints Flow 'Add/Manage GitHub Account' btn
As per Create App From Repo flow, I added the 'Add/ Manage' options to
the Blueprint flow. Added screenshots for both of these correctly, and
updated/added how this behaviour is described. Also updated a Slack
Success Toast image and reference, a GitProvider Tiles Image and title,
a Blueprint image (including move from ACF to ACM in descriptions,
handled in GAIA-3594), and the deletion of some unused clones on some of
the images mentioned.

✅ Closes: GAIA-3582